DEV: Move service tests into modules
